+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Special thanks to the following contributors: @B1gG, @codac, @ezeholz, @khassel,
 - Added custom-properties for colors and fonts for improved styling experience, see `custom.css.sample` file
 - Added custom-properties for gaps around body and between modules
 - Added test case for recurring calendar events
+- If (new) param `url` is provided in module config a missing module will be cloned (and installed) on start
 
 ### Updated